Started building tests
