--- InputHandler/InputHandler/includes/Test.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "ecs.h"
-#include "Test.hpp"
