+++ InputHandler/InputHandler/json/playerMode.json
@@ -1,1 +1,1 @@
-[{"Button":"0","MapTileX":"1","MapTileY":"2","Name":"Player1"},{"Button":"1","MapTileX":"0","MapTileY":"0","Name":"Player2"},{"Button":"1","MapTileX":"0","MapTileY":"0","Name":"Player3"},{"Button":"3","MapTileX":"5","MapTileY":"5","Name":"Player4"}]
+[{"Button":"0","MapTileX":"1","MapTileY":"2","Name":"Player1"},{"Button":"1","MapTileX":"0","MapTileY":"0","Name":"Player2"},{"Button":"2","MapTileX":"0","MapTileY":"0","Name":"Player3"},{"Button":"3","MapTileX":"5","MapTileY":"5","Name":"Player4"}]
